shotdetection/requirements.txt
ffmpeg-python==0.2.0 
grpcio==1.35.0
grpcio-tools==1.35.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
numpy==1.18.4
PyYAML==5.3.1
scikit-image==0.17.2